Added borderWidth, and borderColor to roundRect
Now the roundedRect PageMenuCursor has the ability to have define the borderWidth, and borderColor of the cursor. They are set to optional just in case someone does not require them. It would have been nice to define them as nil by default not to have any API breaking change; however, that is not possible with a Tuple.